requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
lxml>=4.9.0

# Deepseek API (OpenAI-compatible client)
openai>=1.0.0

# Telegram Bot
python-telegram-bot>=21.0

# Testing dependencies
pytest>=8.0.0
pytest-mock>=3.12.0
responses>=0.25.0
playwright>=1.40.0
